Added state machine for potion power-ups, + other changes
- Made a state machine that switches to good or bad potions depending on the trust and greed meters, and randomly picks a stat to boost.
- Converted the item spawner to work as individual object instances rather than a gamemanager spawning randomly.
- Set SceneSwitch to require a number of gatekeys before the level switch activates (will test soon).
- To Do: the player spawning from last checkpoint after choosing "continue" on the game over scene.

+++ ShaytanKids Project/Assets/Scripts/PlayerScripts/PlayerMovement.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class PlayerMovement : MonoBehaviour
 {
     public float currentMoveSpeed;
-    public int maxMoveSpeed;
+    public float maxMoveSpeed;
     public bool moveLeft;
     public bool moveRight;
     public int horizontalInput;
@@ -14,7 +14,7 @@ public class PlayerMovement : MonoBehaviour
     /*public bool shouldSprint;
     public int sprintSpeed;
 */
-    public int jumpHeight;
+    public float jumpHeight;
     public int jumpCount;
     public bool isGrounded;
     public bool canJump;
